Change to S-model with pre-train
Move Channel layer class def to models.py module, with titanic LR
Add the mpl_ kwargs in Scenario options
