+++ mplc/contributivity.py
@@ -100,15 +100,17 @@ class Contributivity:
             small_partners_list = np.array([self.scenario.partners_list[i] for i in subset])
             if len(small_partners_list) > 1:
                 mpl = self.scenario._multi_partner_learning_approach(self.scenario,
-                                                                     partners_list=small_partners_list,
-                                                                     is_early_stopping=True,
-                                                                     save_folder=None,
-                                                                     )
+                                                                    partners_list=small_partners_list,
+                                                                    is_early_stopping=True,
+                                                                    save_folder=None,,
+                                                                    **self.scenario.mpl_kwargs
+                                                                    )
             else:
                 mpl = multi_partner_learning.SinglePartnerLearning(self.scenario,
                                                                    partner=small_partners_list[0],
                                                                    is_early_stopping=True,
                                                                    save_folder=None,
+                                                                   **self.scenario.mpl_kwargs
                                                                    )
             mpl.fit()
             self.charac_fct_values[tuple(subset)] = mpl.history.score
@@ -951,7 +953,8 @@ class Contributivity:
             is_early_stopping=False,
             init_model_from="random_initialization",
             use_saved_weights=False,
-            custom_name='PVRL'
+            custom_name='PVRL',
+            **self.scenario.mpl_kwargs
         )
         full_partners_list = mpl.partners_list  # this list must be a list of PartnerMpl
         initial_model = mpl.build_model()
