+++ mplc/dataset.py
@@ -17,8 +17,6 @@ from joblib import dump, load
 from librosa import load as wav_load
 from librosa.feature import mfcc
 from loguru import logger
-from sklearn.linear_model import LogisticRegression as skLR
-from sklearn.metrics import log_loss
 from sklearn.model_selection import train_test_split
 from tensorflow.keras.datasets import cifar10, mnist, imdb
 from tensorflow.keras.layers import Activation
@@ -32,6 +30,7 @@ from tensorflow.keras.preprocessing import sequence
 from tensorflow.keras.utils import to_categorical
 
 from . import constants
+from .models import LogisticRegression
 
 
 class Dataset(ABC):
@@ -309,7 +308,7 @@ class Titanic(Dataset):
     def generate_new_model(self):
         """Return a LogisticRegression Classifier"""
 
-        clf = self.LogisticRegression()
+        clf = LogisticRegression()
         clf.classes_ = np.array([0, 1])
         clf.metrics_names = ["loss", "accuracy"]  # Mimic Keras's
         return clf
@@ -326,79 +325,6 @@ class Titanic(Dataset):
     @staticmethod
     def train_test_split_global(x, y):
         return train_test_split(x, y, test_size=0.1, random_state=42)
-
-    class LogisticRegression(skLR):
-        def __init__(self):
-            super(Titanic.LogisticRegression, self).__init__(max_iter=10000, warm_start=1, random_state=0)
-            self.coef_ = None
-            self.intercept_ = None
-
-        def fit(self, x_train, y_train, batch_size, validation_data, epochs=1, verbose=False):
-            history = super(Titanic.LogisticRegression, self).fit(x_train, y_train)
-            [loss, acc] = self.evaluate(x_train, y_train)
-            [val_loss, val_acc] = self.evaluate(*validation_data)
-            # Mimic Keras' history
-            history.history = {
-                'loss': [loss],
-                'accuracy': [acc],
-                'val_loss': [val_loss],
-                'val_accuracy': [val_acc]
-            }
-
-            return history
-
-        def evaluate(self, x_eval, y_eval, **kwargs):
-            if self.coef_ is None:
-                model_evaluation = [0] * 2
-            else:
-                loss = log_loss(y_eval, self.predict(x_eval))  # mimic keras model evaluation
-                accuracy = self.score(x_eval, y_eval)
-                model_evaluation = [loss, accuracy]
-
-            return model_evaluation
-
-        def save_weights(self, path):
-            if self.coef_ is None:
-                raise ValueError(
-                    'Coef and intercept are set to None, it seems the model has not been fit properly.')
-            if '.h5' in path:
-                logger.debug('Automatically switch file format from .h5 to .npy')
-                path.replace('.h5', '.npy')
-            np.save(path, self.get_weights())
-
-        def load_weights(self, path):
-            if '.h5' in path:
-                logger.debug('Automatically switch file format from .h5 to .npy')
-                path.replace('.h5', '.npy')
-            weights = load(path)
-            self.set_weights(weights)
-
-        def get_weights(self):
-            if self.coef_ is None:
-                return None
-            else:
-                return np.concatenate((self.coef_, self.intercept_.reshape(1, 1)), axis=1)
-
-        def set_weights(self, weights):
-            if weights is None:
-                self.coef_ = None
-                self.intercept_ = None
-            else:
-                self.coef_ = np.array(weights[0][:-1]).reshape(1, -1)
-                self.intercept_ = np.array(weights[0][-1]).reshape(1)
-
-        def save_model(self, path):
-            if '.h5' in path:
-                logger.debug('Automatically switch file format from .h5 to .joblib')
-                path.replace('.h5', '.joblib')
-            dump(self, path)
-
-        @staticmethod
-        def load_model(path):
-            if '.h5' in path:
-                logger.debug('Automatically switch file format from .h5 to .joblib')
-                path.replace('.h5', '.joblib')
-            return load(path)
 
 
 class Mnist(Dataset):
